==> src/Organoids/data_preprocesing_and_integration/03_Integration2_annotation_Org.Rmd <==
---
title: "03_Integration2_annotation_Org"
output: html_document
date: "2023-03-07"
# Former 04_annotation_auto.Rmd
---

<!--
#-------------------------------------------------------------------------------
# This script performs several steps as part of the annotation process:
# - Load and process in vivo reference datasets
# - Integrates the reference data using the SCTransform (SCT) approach
# - Performs organoid cell annotation using in vivo epithelium 
#   and manually approach using canonical markers
# - Filters out JunB cells
# - Integrate organoid data again

# Author: Carlos Simon Foundation (Raul PM @supermegamio, Diego AC @diegoamoros)
# Date: 2023-03-07
#-------------------------------------------------------------------------------
# Former 02_integration_Bender_Org.Rmd
-->


```{r}
library(Seurat)
library(SeuratDisk)
library(SeuratObject)
library(ggplot2)
library(scales)
library(ggpubr)
library(patchwork)
library(reshape2)
library(yaml)
library(this.path)
options(bitmapType = "cairo")


## Load config.yml file
find_config <- function(names = c("config.yaml", "config.yml")) {
  currrent.path <- dirname(this.path::this.path())
  dirs <- unique(c(currrent.path, dirname(currrent.path)))
  for (d in dirs) {
    for (nm in names) {
      p <- file.path(d, nm)
      if (file.exists(p))
        return(normalizePath(p, winslash = "/", mustWork = TRUE))
    }
  }
  stop("config.yaml|config.yml not found: ",
       paste(dirs, collapse = " o "), call. = FALSE)
}

config <- yaml.load_file(find_config())


## Count matrices processed and filtered (h5seurat format)
input_folder <- config$processed_data$dir

## H5seurat object with all data ans metadata
output_folder <- config$integration_3_data$dir
dir.create(path = output_folder, recursive = TRUE, showWarnings = FALSE)

qc_dir <- config$integration_3_data$qc_dir
dir.create(path = qc_dir, recursive = TRUE, showWarnings = FALSE)


# Load input Seurat Object
sc_AS_org <- LoadH5Seurat(file = file.path(config$integration_2_data$dir,
                                  config$integration_2_data$file))
```





Load and process reference data

```{r}
# WOI reference
sc_pre_woi <- LoadH5Seurat(file = config$integration_3_data$file_ref)
table(sc_pre_woi$principal_cell_types)

sc_pre_woi_Epi <- sc_pre_woi[,which(sc_pre_woi$principal_cell_types %in% 
                      c("AS-Epithelium", "Epithelium","Ciliated Epithelium"))] 
rm(sc_pre_woi)
sc_pre_woi_Epi$cell_type_zooming_Epi <- sc_pre_woi_Epi$cell_type_zooming_v5
table(sc_pre_woi_Epi$cell_type_zooming_Epi)
sc_pre_woi_Epi$Group_epi <- sc_pre_woi_Epi$Group_final

# Asherman in vivo reference
sc_pre_post <- LoadH5Seurat(file = config$integration_3_data$file_ref_as)

sc_pre_post <- sc_pre_post[,which(sc_pre_post$Treatment_stage == "B-Post")]
gc()
sc_pre_post$Treatment_stage <- gsub(pattern = "B-Post", 
                                    replacement = "AS-Post", 
                                    x = sc_pre_post$Treatment_stage)
table(sc_pre_post$cell_type_zoom_Final_v3)
sc_pre_post$cell_type_zooming_Epi <- sc_pre_post$cell_type_zoom_Final_v3
sc_pre_post$Group_epi <- sc_pre_post$Treatment_stage

```

Merge dataset

```{r}
sc_pre_woi_Epi <- DietSeurat(sc_pre_woi_Epi, assays = "RNA")
sc_pre_post <- DietSeurat(sc_pre_post, assays = "RNA")
gc()

sc_AS_epi <- merge(sc_pre_post, sc_pre_woi_Epi)
sc_AS_epi <- NormalizeData(sc_AS_epi)
sc_AS_epi <- FindVariableFeatures(sc_AS_epi, nfeatures = 4000)
sc_AS_epi$cell_type_zooming_Epi <- gsub(pattern = "Epi-Lumenal", 
                                        replacement = "Epi-Luminal", 
                                        x = sc_AS_epi$cell_type_zooming_Epi)
table(sc_AS_epi$Group_epi)

sc_AS_epi$Group_epi <- gsub(pattern = "WOI WOI Control", 
                            replacement = "WOI Control", 
                            x = sc_AS_epi$Group_epi)
sc_AS_epi$Group_epi[which(sc_AS_epi$Group_epi == "AS")] <- "AS-Pre"
table(sc_AS_epi$Group_epi)


```


```{r}
sc_AS_lineage_list <- SplitObject(sc_AS_epi, split.by = "Group_epi")
rm(sc_AS_epi)
gc()
# normalize and identify variable features for each dataset independently
  sc_AS_lineage_list <- lapply(
    X = sc_AS_lineage_list, 
    function(x){SCTransform(object = x, 
                            vars.to.regress = c("mitoRatio","Phase"), 
                            conserve.memory = T, variable.features.n = 1000)})

# select features that are repeatedly variable across datasets for integration
  features <- SelectIntegrationFeatures(
    object.list = sc_AS_lineage_list, nfeatures = 1000)
  sc_AS_lineage_list <- PrepSCTIntegration(
    object.list = sc_AS_lineage_list, anchor.features = features)
  sc_AS.anchors <- FindIntegrationAnchors(
    object.list = sc_AS_lineage_list, normalization.method = "SCT", 
    anchor.features = features)
  rm(sc_AS_lineage_list)
  gc()
  sc_AS_lineage_list_integrated <- IntegrateData(
    anchorset = sc_AS.anchors, normalization.method = "SCT")
  
  # Run the standard workflow for visualization and clustering
  sc_AS_lineage_list_integrated <- RunPCA(sc_AS_lineage_list_integrated, 
                                          npcs = 30,  verbose = FALSE)
  sc_AS_lineage_list_integrated <- RunUMAP(sc_AS_lineage_list_integrated, 
                                           reduction = "pca", dims = 1:30)
  sc_AS_lineage_list_integrated <- FindNeighbors(sc_AS_lineage_list_integrated, 
                                                 reduction = "pca", dims = 1:30)
  sc_AS_lineage_list_integrated <- FindClusters(sc_AS_lineage_list_integrated,
                                                resolution = 0.5)


DimPlot(object = sc_AS_lineage_list_integrated, 
        group.by = "cell_type_zooming_Epi", label = T, split.by = "Group_epi")

```

Cell annotation from in vivo epithelium into in vitro organoids

```{r}
DefaultAssay(sc_AS_org) <- "integrated"
sc_epi_org.anchors <- FindTransferAnchors(
  reference = sc_AS_lineage_list_integrated, 
  query = sc_AS_org, dims = 1:30, reference.reduction = "pca")

predictions_Epi_AS <- TransferData(
  anchorset = sc_epi_org.anchors, 
  refdata = sc_AS_lineage_list_integrated$cell_type_zooming_Epi,
    dims = 1:30)

predictions_Epi_AS$final_id <- predictions_Epi_AS$predicted.id
predictions_Epi_AS$final_id[
  which(predictions_Epi_AS$prediction.score.max < 0.6)
  ] <- "weak_assigment" 

sc_AS_org$cell_type_predict <- predictions_Epi_AS$final_id

DimPlot(object = sc_AS_org, group.by = "cell_type_predict", 
        label = T, split.by = "Group_simple")
```

# Manual cell id

```{r}
DimPlot(object = sc_AS_org, group.by = "seurat_clusters", label = T)
table(sc_AS_org$Patient, sc_AS_org$Group_simple)
sc_AS_org$Group_simple <- factor(sc_AS_org$Group_simple, 
                                 levels = c("AS-Pre", "AS-Post",
                                            "LH+7", "LH+8"))

cell_clusters_markers <- 
  read.csv(file = file.path(config$integration_2_data$qc_dir,
                           "02_organodis_0.5_res_markers.csv"))

cannonical_markers <- read.csv(config$integration_3_data$file_markers)

dotplot_cannonical <- DotPlot(object = sc_AS_org, assay = "RNA", 
                              features = cannonical_markers$Gene,
                              cols = "RdYlBu") + coord_flip()

```




```{r}
library(DelayedArray)
library(ComplexHeatmap)
library(dplyr)

zscore_sc <- function(seurat_obj) {
  
  
  counts <- DelayedArray::DelayedArray(seurat_obj@assays$RNA@data)
  counts <- (counts - DelayedMatrixStats::rowMeans2(counts)) / 
    DelayedMatrixStats::rowSds(counts)
  counts[base::is.nan(counts)] <- 0
  return(counts)
}

cell_clusters_markers %>%
    group_by(cluster) %>%
    top_n(n = 5, wt = avg_log2FC) -> top5


zscore_seurat <- zscore_sc(sc_AS_org)
cluster_annot <- sc_AS_org@meta.data
zscore_seurat <- zscore_seurat[unique(top5$gene),]
quantiles_matrix <- quantile(zscore_seurat, c(0.1, 0.95))

col_fun = circlize::colorRamp2(c(quantiles_matrix[1],0,quantiles_matrix[2]),
                               c("#FF00FF", "black", "#FFFF00"))

heatmap <- Heatmap(as.matrix(zscore_seurat), name = "z-score",  
        column_split = factor(cluster_annot[,"seurat_clusters"]),
        cluster_columns = TRUE,
        show_column_dend = TRUE,
        cluster_column_slices = TRUE,
        column_title_gp = gpar(fontsize = 8),
        column_gap = unit(0.5, "mm"),
        cluster_rows = TRUE,
        show_row_dend = FALSE,
        col = col_fun,
        row_names_gp = gpar(fontsize = 4),
        column_title_rot = 90,
        top_annotation = HeatmapAnnotation(
          foo = anno_block(labels_gp = gpar(fontsize = 5),
                           gp = gpar(fill = scales::hue_pal()(9),
                                     fontsize = 5))),
        show_column_names = FALSE,
        use_raster = TRUE,
        raster_quality = 4)

svg(
  filename = file.path(qc_dir,"03-heatmap_test.svg"),
  width = 16, height = 10)
draw(heatmap)
dev.off()

```


10 12 14: Glandular Secretory
15: Ciliated
11: Remove-JuNB
9: SAA1/SAA2+ 
5: Cycling organoids
0,7,8: MMP7 proliferative 
13: KRT17+ 
4: Glandular

```{r}
epi_clust <- c("Proliferative MMP7+","Glandular Secretory-1","SAA1/SAA2+",
               "SAA1/SAA2+","Glandular","Cycling Org","SAA1/SAA2+",
               "Proliferative MMP7+","Proliferative MMP7+","SAA1/SAA2+",
               "Glandular Secretory","JunB","Glandular Secretory","KRT17+",
               "Glandular Secretory","Ciliated")
epi_clust_id <- c(0:15)

sc_AS_org$cell_type_manual_org <- plyr::mapvalues(x = sc_AS_org$seurat_clusters,
                                                  to = epi_clust, 
                                                  from = epi_clust_id)
DimPlot(object = sc_AS_org, group.by = "cell_type_manual_org", label = T)

```

```{r}
sc_AS_org_filt2 <- sc_AS_org[,which(sc_AS_org$cell_type_manual_org != "JunB")]

vln_fil <- VlnPlot(object = sc_AS_org_filt2, features = "nFeature_RNA", 
                   group.by = "orig.ident", log = T, pt.size = 0) + NoLegend()
mitoRatio <- VlnPlot(object = sc_AS_org_filt2, features = "mitoRatio", 
                     group.by = "orig.ident", log = F, pt.size = 0) + NoLegend()
bar_plot_sample <- dittoSeq::dittoBarPlot(object = sc_AS_org_filt2, 
                                          group.by = "orig.ident", 
                                          var = "cell_type_manual_org")
bar_plot_sample_count <- dittoSeq::dittoBarPlot(object = sc_AS_org_filt2, 
                                                group.by = "orig.ident",
                                                var = "cell_type_manual_org", 
                                                scale = "count")




table(sc_AS_org_filt2$orig.ident)
```





Second integration without JunB



```{r}
sc_AS_lineage_list <- SplitObject(sc_AS_org_filt2, split.by = "Group_simple")

# normalize and identify variable features for each dataset independently
  sc_AS_lineage_list <- lapply(X = sc_AS_lineage_list, function(x){SCTransform(
    object = x, vars.to.regress = c("mitoRatio","Phase"), 
    conserve.memory = T, variable.features.n = 750)})

# select features that are repeatedly variable across datasets for integration
  features <- SelectIntegrationFeatures(object.list = sc_AS_lineage_list, 
                                        nfeatures = 750)
  sc_AS_lineage_list <- PrepSCTIntegration(object.list = sc_AS_lineage_list, 
                                           anchor.features = features)
  sc_AS.anchors <- FindIntegrationAnchors(object.list = sc_AS_lineage_list,
                                          normalization.method = "SCT", 
                                          anchor.features = features)
  sc_AS_lineage_list_integrated <- IntegrateData(anchorset = sc_AS.anchors,
                                                 normalization.method = "SCT")
  # Run the standard workflow for visualization and clustering
  sc_AS_lineage_list_integrated <- RunPCA(sc_AS_lineage_list_integrated, 
                                          npcs = 30,  verbose = FALSE)
  sc_AS_lineage_list_integrated <- RunUMAP(sc_AS_lineage_list_integrated, 
                                           reduction = "pca", dims = 1:30)
  sc_AS_lineage_list_integrated <- FindNeighbors(sc_AS_lineage_list_integrated,
                                                 reduction = "pca", dims = 1:30)
  sc_AS_lineage_list_integrated <- FindClusters(sc_AS_lineage_list_integrated, 
                                                resolution = 0.5)


```

Find clusters

relabeling metadata

```{r}
metadata_org <- read.csv(file = config$metadata$file)
group <- plyr::mapvalues(
  x = sc_AS_lineage_list_integrated$orig.ident,
  to = metadata_org$Group, from = metadata_org$Sample)
patient <- plyr::mapvalues(
  x = sc_AS_lineage_list_integrated$orig.ident, 
  to = metadata_org$Patient, from = metadata_org$Sample)

sc_AS_lineage_list_integrated <- AddMetaData(
  object = sc_AS_lineage_list_integrated, metadata = group,
  col.name = "Group")

sc_AS_lineage_list_integrated <- AddMetaData(
  object = sc_AS_lineage_list_integrated, metadata = patient, 
  col.name = "Patient")
```


```{r}
orig_id <- sc_AS_lineage_list_integrated$orig.ident
orig_id <- gsub(
  pattern = "run_count__0413__RWD-23F00[0-9]{2}_|_CeBender_filtered",
  replacement = "", x = orig_id)
sc_AS_lineage_list_integrated$orig_id_filt <- orig_id

bar_plot_sample_count <- dittoSeq::dittoBarPlot(
  object = sc_AS_lineage_list_integrated, group.by = "orig_id_filt", 
  var = "cell_type_manual_org", scale = "count")
bar_plot_sample <- dittoSeq::dittoBarPlot(
  object = sc_AS_lineage_list_integrated, group.by = "orig_id_filt", 
  var = "cell_type_manual_org")
multi_bar <- ggarrange(bar_plot_sample, bar_plot_sample_count,
                       ncol = 2, common.legend = T)


## Save processed Seurat object
SaveH5Seurat(object = sc_AS_lineage_list_integrated, 
             filename = file.path(output_folder,
                                  config$integration_3_data$file))

```
