==> src/Organoids/data_preprocesing_and_integration/05_Integration4_annotation_Org.rmd <==
---
title: "05_Integration4_annotation_Org"
output: html_document
date: "2023-03-08"
---

<!--
#-------------------------------------------------------------------------------
# Final integration. This script performs several steps:
# - Filters out cells with less of 750 genes detected 
# - Performs the integration pipeline again using SCT approach
# - Dimensional reduction and clusterization
# Extended and cell type markers adapted from Santamaria X et al. Nat Commun 14:5890
#
# Author: Carlos Simon Foundation (Raul PM @supermegamio, Diego AC @diegoamoros)
# Date: 2023-03-07
#-------------------------------------------------------------------------------
# Former 06_750_org
-->

```{r}
library(Seurat)
library(SeuratDisk)
library(SeuratObject)
library(ggplot2)
library(scales)
library(ggpubr)
library(patchwork)
library(reshape2)
library(DelayedArray)
library(ComplexHeatmap)
library(dplyr)
library(yaml)
library(this.path)
options(bitmapType="cairo")


## Load config.yml file
find_config <- function(names = c("config.yaml", "config.yml")) {
  currrent.path <- dirname(this.path::this.path())
  dirs <- unique(c(currrent.path, dirname(currrent.path)))
  for (d in dirs) {
    for (nm in names) {
      p <- file.path(d, nm)
      if (file.exists(p))
        return(normalizePath(p, winslash = "/", mustWork = TRUE))
    }
  }
  stop("config.yaml|config.yml not found: ",
       paste(dirs, collapse = " o "), call. = FALSE)
}

config <- yaml.load_file(find_config())



## ouput H5seurat object folder
output_folder <- config$integration_5_data$dir
dir.create(path = output_folder, recursive = TRUE, showWarnings = FALSE)

qc_dir <- config$integration_5_data$qc_dir
dir.create(path = qc_dir, recursive = TRUE, showWarnings = FALSE)


# Load input Seurat Object
sc_AS_org <- LoadH5Seurat(
  file = file.path(config$integration_4_data$dir,
                   config$integration_4_data$file))

annotations <- readRDS(
      file = config$integration_5_data$file_annotations)
```


Only cells with more of 750 genes 

```{r}

#Filter out cells with less of 750 genes detected 
sc_AS_org$less_750 <- sc_AS_org$nFeature_RNA < 750
sc_AS_org <- sc_AS_org[,which(sc_AS_org$less_750 == FALSE)]

# Integration pipeline again using SCT approach
sc_AS_lineage_list <- SplitObject(sc_AS_org, split.by = "Group")

# normalize and identify variable features for each dataset independently
  sc_AS_lineage_list <- lapply(
    X = sc_AS_lineage_list, 
    function(x){SCTransform(
      object = x, 
      vars.to.regress = c("mitoRatio","Phase"), 
      conserve.memory = T, variable.features.n = 750)})

# select features that are repeatedly variable across datasets for integration
  features <- SelectIntegrationFeatures(
    object.list = sc_AS_lineage_list, nfeatures = 750)
  sc_AS_lineage_list <- PrepSCTIntegration(
    object.list = sc_AS_lineage_list, anchor.features = features)
  sc_AS.anchors <- FindIntegrationAnchors(
    object.list = sc_AS_lineage_list, normalization.method = "SCT", 
    anchor.features = features)
  sc_AS_lineage_list_integrated <- IntegrateData(
    anchorset = sc_AS.anchors, normalization.method = "SCT")
  
  # Run the standard workflow for visualization and clustering
  sc_AS_lineage_list_integrated <- RunPCA(sc_AS_lineage_list_integrated, 
                                          npcs = 30,  verbose = FALSE)
  sc_AS_lineage_list_integrated <- RunUMAP(sc_AS_lineage_list_integrated, 
                                           reduction = "pca", dims = 1:30)
  sc_AS_lineage_list_integrated <- FindNeighbors(sc_AS_lineage_list_integrated, 
                                                 reduction = "pca", dims = 1:30)
  sc_AS_lineage_list_integrated <- FindClusters(sc_AS_lineage_list_integrated, 
                                                resolution = 0.5)

```

```{r}

umap <- DimPlot(sc_AS_lineage_list_integrated)
ggsave(filename = file.path(qc_dir,"umap.svg"), 
       plot = umap, device = "svg", width = 8, height = 8)



```

```{r}
DefaultAssay(sc_AS_lineage_list_integrated) <- "RNA"
sc_AS_lineage_list_integrated <- NormalizeData(sc_AS_lineage_list_integrated)


sc_AS_lineage_list_integrated <- SetIdent(
  object = sc_AS_lineage_list_integrated, value = "seurat_clusters")
sc_markers_seurat <- FindAllMarkers(
  object = sc_AS_lineage_list_integrated, assay = "RNA", min.pct = 0.3)

sc_AS_lineage_list_integrated <- SetIdent(
  object = sc_AS_lineage_list_integrated, value = "cell_type_manual_org")
sc_markers_mainpop <- FindAllMarkers(
  object = sc_AS_lineage_list_integrated, assay = "RNA", min.pct = 0.3)

sc_markers_mainpop <- sc_markers_mainpop %>%
    left_join(y = unique(annotations[, c("gene_name", "description")]),
               by = c("gene" = "gene_name"))

sc_markers_seurat <- sc_markers_seurat %>%
    left_join(y = unique(annotations[, c("gene_name", "description")]),
               by = c("gene" = "gene_name"))
```




```{r}


zscore_sc <- function(seurat_obj) {
  
  
  counts <- DelayedArray::DelayedArray(seurat_obj@assays$RNA@data)
  counts <- (counts - DelayedMatrixStats::rowMeans2(counts)) / 
    DelayedMatrixStats::rowSds(counts)
  counts[base::is.nan(counts)] <- 0
  return(counts)
}

sc_markers_seurat %>%
    group_by(cluster) %>%
    top_n(n = 5, wt = avg_log2FC) -> top5


zscore_seurat <- zscore_sc(sc_AS_lineage_list_integrated)
cluster_annot <- sc_AS_lineage_list_integrated@meta.data
zscore_seurat <- zscore_seurat[unique(top5$gene),]
quantiles_matrix <- quantile(zscore_seurat, c(0.1, 0.95))

col_fun = circlize::colorRamp2(c(quantiles_matrix[1],0,quantiles_matrix[2]), 
                               c("#FF00FF", "black", "#FFFF00"))

heatmap <- Heatmap(as.matrix(zscore_seurat), name = "z-score",  
        column_split = factor(cluster_annot[,"seurat_clusters"]),
        cluster_columns = TRUE,
        show_column_dend = T,
        cluster_column_slices = TRUE,
        column_title_gp = gpar(fontsize = 8),
        column_gap = unit(0.5, "mm"),
        cluster_rows = TRUE,
        show_row_dend = FALSE,
        col = col_fun,
        row_names_gp = gpar(fontsize = 4),
        column_title_rot = 90,
        top_annotation = HeatmapAnnotation(
          foo = anno_block(labels_gp = gpar(fontsize = 5),
                          gp = gpar(fill = scales::hue_pal()(9),fontsize = 5))),
        show_column_names = FALSE,
        use_raster = TRUE,
        raster_quality = 4)

svg(
  filename = file.path(qc_dir,"heatmap_org_seurat_filt.svg"), 
  width = 16, height = 10)
 draw(heatmap)
 dev.off()

```

```{r}
DefaultAssay(sc_AS_lineage_list_integrated) <- "integrated"
sc_AS_lineage_list_integrated <- FindClusters(sc_AS_lineage_list_integrated, 
                                              resolution = 2)
DefaultAssay(sc_AS_lineage_list_integrated) <- "RNA"
cluster_34 <- FindMarkers(object = sc_AS_lineage_list_integrated, 
                          ident.1 = "34", group.by = "seurat_clusters")

#clsuter 34 are doublets 
sc_AS_lineage_list_integrated <- sc_AS_lineage_list_integrated[,which(
  sc_AS_lineage_list_integrated$integrated_snn_res.2 != 34)]
```

```{r}
umap_seurat_clusters <- DimPlot(object = sc_AS_lineage_list_integrated, 
                                group.by = "seurat_clusters", label = T)
umap_old_id <- DimPlot(object = sc_AS_lineage_list_integrated, 
                       group.by = "cell_type_manual_org", label = T)
multi_umap <- ggarrange(umap_seurat_clusters, umap_old_id)
ggsave(filename = file.path(qc_dir,"05_multi_umap.svg"), 
       device = "svg", plot = multi_umap, width = 10, height = 5)

epi_cluster_final <- c("Proliferative MMP7+","Glandular Secretory",
                       "EEO SAA1/SAA2","Glandular","EEO SAA1/SAA2",
                       "Cycling Org","Proliferative MMP7+",
                       "Glandular Secretory","Glandular Secretory",
                       "Proliferative MMP7+","KRT17+","Ciliated",
                       "EEO SAA1/SAA2","Glandular Secretory")
epi_clsuter_id <- c(0:13)

sc_AS_lineage_list_integrated$cell_type_manual_org_v2 <- plyr::mapvalues(
  x = sc_AS_lineage_list_integrated$integrated_snn_res.0.5, 
  to = epi_cluster_final, from = epi_clsuter_id)
DimPlot(object = sc_AS_lineage_list_integrated, 
        group.by = "cell_type_manual_org_v2", label = T)
```


```{r}
proliferative_color <- "#E69F00"
glandular_secretory_color <- "#0064A4"
eeo_color <- "#CC79A7"
mmp7_color <- "#56B4E9"
cycling_org_color <- "#009E73"
krt17_color <- "#F0E442"
glandular_color <- "#6688CC"
ciliated_color <- "#D55E00"

saa_color <- "#000000"


cell_colors <- c(proliferative_color, glandular_secretory_color, 
                 eeo_color, glandular_color,
            cycling_org_color, krt17_color, ciliated_color)
sc_AS_lineage_list_integrated$Group <- gsub(
  pattern = "LH\\+7", replacement = "Control", 
  x = sc_AS_lineage_list_integrated$Group)

sc_AS_lineage_list_integrated$Group <- 
  factor(sc_AS_lineage_list_integrated$Group, 
         levels = c("AS-Pre","AS-Post","Control"))

umap1_cell <- DimPlot(object = sc_AS_lineage_list_integrated, 
                      group.by = "cell_type_manual_org_v2",
                      cols = cell_colors, label = T)

umap1_cell <- DimPlot(object = sc_AS_lineage_list_integrated, 
                      group.by = "cell_type_manual_org_v2", 
                      cols = cell_colors, label = T, 
                      split.by = "Group", raster = T, raster.dpi = c(1080,1080))

umap1_cell_raster_f <- DimPlot(object = sc_AS_lineage_list_integrated, 
                               group.by = "cell_type_manual_org_v2", 
                               cols = cell_colors, label = T, 
                               split.by = "Group", raster = F)

ggsave(filename = file.path(qc_dir,"05-fig_A_umap.svg"), 
       plot = umap1_cell, device = "svg", width = 16, height = 8, bg = "white")
ggsave(filename = file.path(qc_dir,"05-fig_A_umap_raster_F.svg"),
       plot = umap1_cell_raster_f, device = "svg", width = 14, 
       height = 6, bg = "white")

## Cell number across patients

bar_plot_cell_types <- dittoSeq::dittoBarPlot(
  object = sc_AS_lineage_list_integrated, var = "cell_type_manual_org_v2", 
  group.by = "orig_id_filt", color.panel = cell_colors, 
  split.by = "Group", split.adjust = list(scale = "free"), 
  scale = "count") + xlab("") + ggtitle("Cell type distribution across samples")
ggsave(filename = file.path(qc_dir,"05-fig_A_bar_supp.svg"), 
  plot = bar_plot_cell_types, device = "svg", 
  width = 12.5, height = 8, bg = "white")

## Cell types across group

bar_plot_cell_types_g <- dittoSeq::dittoBarPlot(
  object = sc_AS_lineage_list_integrated, 
  var = "cell_type_manual_org_v2", group.by = "Group", 
  color.panel = cell_colors, split.adjust = 
    list(scale = "free")) + xlab("") + 
  ggtitle("Cell type distribution across groups")
ggsave(filename = file.path(qc_dir,"05-fig_B_bar_supp.svg"), 
  plot = bar_plot_cell_types_g, device = "svg", 
  width = 6, height = 8, bg = "white")

```

#Final cell types

```{r}

sc_AS_lineage_list_integrated$cell_type_manual_org_v3 <- 
  as.character(sc_AS_lineage_list_integrated$cell_type_manual_org_v2)

sc_AS_lineage_list_integrated$cell_type_manual_org_v3[
  sc_AS_lineage_list_integrated$cell_type_manual_org_v3 ==
    "EEO SAA1/SAA2" ] <- "Poliferative EMT"

sc_AS_lineage_list_integrated$cell_type_manual_org_v3[
  sc_AS_lineage_list_integrated$cell_type_manual_org_v3 ==
    "Proliferative MMP7+" ] <- "Proliferative KRT19"


umap1_cell <- DimPlot(object = sc_AS_lineage_list_integrated, 
                      group.by = "cell_type_manual_org_v3", 
                      cols = cell_colors, label = T, pt.size = 4, pt.size = 2)

umap1_cell <- DimPlot(object = sc_AS_lineage_list_integrated, 
                      group.by = "cell_type_manual_org_v3", 
                      cols = cell_colors, label = T, split.by = "Group", 
                      raster = T, raster.dpi = c(1080,1080), pt.size = 2)

umap1_cell_raster_f <- DimPlot(object = sc_AS_lineage_list_integrated, 
                               group.by = "cell_type_manual_org_v3", 
                               cols = cell_colors, label = T, 
                               split.by = "Group", raster = F, pt.size = 1)

ggsave(filename = file.path(qc_dir,"05-fig_A_umap_paper.svg"), 
       plot = umap1_cell, device = "svg", width = 16, height = 8, 
       bg = "white")

ggsave(filename = file.path(qc_dir,"05-fig_A_umap_raster_F_paper.svg"), 
  plot = umap1_cell_raster_f, device = "svg", width = 14, height = 6, 
  bg = "white")


## Save processed Seurat object
SaveH5Seurat(object = sc_AS_lineage_list_integrated, 
             filename = file.path(output_folder,
                   config$integration_5_data$file))

```
